refactor: cleanup analyst application fields and refactor shared tests
