+++ app/tests/components/Form/ApplicationForm.test.tsx
@@ -3,12 +3,13 @@ import { graphql } from 'react-relay';
 import compiledQuery, {
   ApplicationFormTestQuery,
 } from '__generated__/ApplicationFormTestQuery.graphql';
-import { fireEvent, screen, within } from '@testing-library/react';
+import { fireEvent, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import mockFormData from 'tests/utils/mockFormData';
 import uiSchema from 'formSchema/uiSchema/uiSchema';
 import { acknowledgementsEnum } from 'formSchema/pages/acknowledgements';
 import ComponentTestingHelper from '../../utils/componentTestingHelper';
+import sharedReviewThemeTests from '../Review/ReviewTheme';
 
 const testQuery = graphql`
   query ApplicationFormTestQuery @relay_test_operation {
@@ -730,440 +731,14 @@ describe('The application form', () => {
     const REVIEW_PAGE_INDEX =
       uiSchema['ui:order'].findIndex((e) => e === 'review') + 1;
 
-    it('should have correct sections', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
+    // Tests for the Review theme that are shared with the review page and analyst application view
+    sharedReviewThemeTests((payload) => {
+      componentTestingHelper.loadQuery(payload);
       componentTestingHelper.renderComponent((data) => ({
         application: data.application,
         pageNumber: REVIEW_PAGE_INDEX,
         query: data.query,
       }));
-
-      screen.getByRole('heading', { name: 'Project information' });
-      screen.getByRole('heading', { name: 'Project area' });
-      screen.getByRole('heading', { name: 'Existing network coverage' });
-      screen.getByRole('heading', { name: 'Budget details' });
-      screen.getByRole('heading', { name: 'Project funding' });
-      screen.getByRole('heading', { name: 'Other funding sources' });
-      screen.getByRole('heading', { name: 'Technological solution' });
-      screen.getByRole('heading', { name: 'Benefits' });
-      screen.getByRole('heading', { name: 'Project planning and management' });
-      screen.getByRole('heading', { name: 'Estimated project employment' });
-      screen.getByRole('heading', { name: 'Template uploads' });
-      screen.getByRole('heading', { name: 'Supporting documents' });
-      screen.getByRole('heading', { name: 'Coverage' });
-      screen.getByRole('heading', { name: 'Organization profile' });
-      screen.getByRole('heading', { name: 'Organization location' });
-      screen.getByRole('heading', { name: 'Organization contact information' });
-    });
-
-    it('should have correct heading styles', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-      const heading = screen.getByRole('heading', {
-        name: 'Project information',
-      });
-
-      const style = window.getComputedStyle(heading);
-
-      expect(style.fontSize).toBe('24px');
-      expect(style.marginBottom).toBe('0px');
-    });
-
-    it('should have correct subheading styles', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-      const subheading = screen.getAllByText(
-        'Amount requested under source:'
-      )[0];
-
-      const style = window.getComputedStyle(subheading);
-
-      expect(style.fontSize).toBe('14px');
-      expect(style.fontWeight).toBe('600');
-      expect(style.padding).toBe('16px');
-      expect(style.margin).toBe('0px');
-    });
-
-    it('should have the correct fields in the Project Area section', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      const section = within(
-        screen.getByRole('heading', { name: 'Project area' }).closest('section')
-      );
-
-      expect(
-        section
-          .getByText(/which zone\(s\) will this project be conducted in\?/i)
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('1');
-
-      expect(
-        section
-          .getByText(
-            /Does your Project span multiple provinces\/territories\?/i
-          )
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent(/yes/i);
-
-      expect(
-        section
-          .getByText(/select the provinces or territories/i)
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent(/Alberta,Northwest Territories/i);
-    });
-
-    it('should have correct fields in the budget details section', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      const section = within(
-        screen
-          .getByRole('heading', { name: 'Budget details' })
-          .closest('section')
-      );
-
-      expect(
-        section
-          .getByText(/Total project cost/i)
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('$1 230');
-
-      expect(
-        section
-          .getByText(/Total eligible cost/i)
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('$1 000');
-    });
-
-    it('should display headings in the estimated employment', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      const section = within(
-        screen
-          .getByRole('heading', { name: 'Estimated project employment' })
-          .closest('section')
-      );
-
-      expect(
-        section.getByText('Estimated direct employees')
-      ).toBeInTheDocument();
-      expect(
-        section.getByText('Estimated contractor labour')
-      ).toBeInTheDocument();
-    });
-
-    it('should display empty red fields in the budget details section when there are errors', () => {
-      componentTestingHelper.loadQuery(mockQueryPayload);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      const section = within(
-        screen
-          .getByRole('heading', { name: 'Budget details' })
-          .closest('section')
-      );
-
-      const errorCell = section
-        .getByText('Total project cost (Template 2 - cell H28)')
-        .closest('tr')
-        .getElementsByTagName('td')[0];
-
-      expect(errorCell).toBeEmptyDOMElement();
-      expect(errorCell).toHaveStyle(
-        'background-color: rgba(248, 214, 203, 0.4)'
-      );
-    });
-
-    it('should have correct fields in the Existing network coverage section', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      const section = within(
-        screen
-          .getByRole('heading', { name: 'Existing network coverage' })
-          .closest('section')
-      );
-
-      expect(
-        section
-          .getByText(/Does the Applicant own Passive Infrastructure/i)
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent(/yes/i);
-    });
-
-    it('should have correct fields in the Project funding  section', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      expect(
-        document.getElementById('root_projectFunding_fundingRequestedCCBC2223')
-      ).toHaveTextContent(`2022-23`);
-
-      expect(
-        document.getElementById('root_projectFunding_fundingRequestedCCBC2324')
-      ).toHaveTextContent(`2023-24`);
-
-      expect(
-        document.getElementById('root_projectFunding_fundingRequestedCCBC2425')
-      ).toHaveTextContent(`2024-25`);
-
-      expect(
-        document.getElementById('root_projectFunding_fundingRequestedCCBC2526')
-      ).toHaveTextContent(`2025-26`);
-
-      expect(
-        document.getElementById('root_projectFunding_fundingRequestedCCBC2627')
-      ).toHaveTextContent(`2026-27`);
-
-      expect(
-        document.getElementById('root_projectFunding_totalFundingRequestedCCBC')
-      ).toHaveTextContent(`Total amount requested under CCBC`);
-
-      expect(
-        document.getElementById(
-          'root_projectFunding_applicationContribution2223'
-        )
-      ).toHaveTextContent(`2022-23`);
-
-      expect(
-        document.getElementById(
-          'root_projectFunding_applicationContribution2324'
-        )
-      ).toHaveTextContent(`2023-24`);
-
-      expect(
-        document.getElementById(
-          'root_projectFunding_applicationContribution2425'
-        )
-      ).toHaveTextContent(`2024-25`);
-
-      expect(
-        document.getElementById(
-          'root_projectFunding_applicationContribution2526'
-        )
-      ).toHaveTextContent(`2025-26`);
-
-      expect(
-        document.getElementById(
-          'root_projectFunding_applicationContribution2627'
-        )
-      ).toHaveTextContent(`2026-27`);
-
-      expect(
-        document.getElementById(
-          'root_projectFunding_totalApplicantContribution'
-        )
-      ).toHaveTextContent(`Total amount Applicant will contribute`);
-    });
-
-    it('should have correct fields in The Technological solution section', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      const section = within(
-        screen
-          .getByRole('heading', { name: /Technological solution/i })
-          .closest('section')
-      );
-
-      expect(
-        section
-          .getByText(/System design/i)
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('design of system');
-
-      expect(
-        section
-          .getByText(/Scalability/i)
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('much scalable');
-
-      expect(
-        section
-          .getByText(/backbone technology/i)
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('Fibre,Satellite');
-
-      expect(
-        section
-          .getByText(/last mile technology/i)
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('Fibre,Fixed wireless');
-    });
-
-    it('should have correct fields in The Project planning section', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      const section = within(
-        screen
-          .getByRole('heading', { name: /Project planning/i })
-          .closest('section')
-      );
-
-      expect(
-        section
-          .getByText('Project Start Date (YYYY/MM/DD)')
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('2022-06-10');
-    });
-
-    it('should have correct fields in The Other funding sources section', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      const section = within(
-        screen
-          .getByRole('heading', { name: 'Other funding sources' })
-          .closest('section')
-      );
-
-      expect(
-        section
-          .getAllByText('2022-23')[0]
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('$1');
-
-      expect(
-        section
-          .getAllByText('2023-24')[0]
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('$2');
-
-      expect(
-        section
-          .getAllByText('2024-25')[0]
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('$3');
-
-      expect(
-        section
-          .getAllByText('2025-26')[0]
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('$4');
-
-      expect(
-        section
-          .getAllByText('2026-27')[0]
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('$5');
-
-      expect(
-        section
-          .getByText(
-            /Total amount requested under Canadian Infrastructure Bank/i
-          )
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent('$15');
-    });
-
-    it('should the correct fields in the Organization location section', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      const section = within(
-        screen
-          .getByRole('heading', { name: 'Organization location' })
-          .closest('section')
-      );
-
-      expect(
-        section
-          .getAllByText(/city/i)[0]
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent(/victoria/i);
-
-      expect(
-        section
-          .getAllByText(/unit number/i)[0]
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent(/1231231/i);
-
-      expect(
-        section
-          .getByText(/Is the mailing address the same as above/i)
-          .closest('tr')
-          .getElementsByTagName('td')[0]
-      ).toHaveTextContent(/no/i);
-    });
-
-    it('should have the correct field and value in the Organization profile section', () => {
-      componentTestingHelper.loadQuery(mockQueryPayloadWithFormData);
-      componentTestingHelper.renderComponent((data) => ({
-        application: data.application,
-        pageNumber: REVIEW_PAGE_INDEX,
-        query: data.query,
-      }));
-
-      expect(document.getElementById('bandNumber-error')).toBeNull();
     });
 
     it('should not display alert box without errors', () => {
